Change room number now support open format
